% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_transformation.R
\name{add_transformation}
\alias{add_transformation}
\title{Add Matrix Transformation}
\usage{
add_transformation(from = matrix(c(1, 0, 0, 1), nrow = 2), m,
  seq_fun = animatrixr::seq_matrix_l, n_frames = 20)
}
\arguments{
\item{from}{Matrix or list of matrices.}

\item{m}{Matrix transformation.}

\item{seq_fun}{The function type for the matrix transformation, `animatrixr::seq_matrix_l`}

\item{n_frames}{Number of frames to create explicitly, defaults to 20.}
}
\value{
List of matrices.
}
\description{
Creates list of matrices. If list is inputted then concatenates onto list of matrices.
}
