% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_matrix.R
\name{animate_matrix}
\alias{animate_matrix}
\title{Animate matrix}
\usage{
animate_matrix(transformations, datasaurus = FALSE,
  return_graph_objects = FALSE, return_static = FALSE)
}
\arguments{
\item{transformations}{List of matrix transformations to apply (length will
determine total number of explicit frames in animation.)}

\item{datasaurus}{Whether to add coordinates of the datasaurus.}

\item{return_graph_objects}{Return the dataframe containing the objects that
would have gone into the animation.}

\item{return_static}{Return the static ggplot graph.}
}
\value{
An animation created by `gganimate`.
}
\description{
Animate matrix
}
