% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatrixr-package.R
\docType{package}
\name{animatrixr}
\alias{animatrixr}
\alias{_PACKAGE}
\alias{animatrixr-package}
\title{Animate 2x2 transformation matrices}
\description{
Toy package in support of my blog post on visualizing simple 2x2
  matrix transformations. Not meant for general use.
}
\keyword{internal}
