% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_transform_coordinates.R
\name{construct_grid}
\alias{construct_grid}
\title{Create coordinates for a grid.}
\usage{
construct_grid(xintercepts = -5:5, yintercepts = -5:5)
}
\arguments{
\item{xintercepts}{Numeric vector, default is -5:5.}

\item{yintercepts}{Numeric vector, default is -5:5.}
}
\value{
Dataframe.
}
\description{
given vectors of x and y intercepts, return a dataframe with columns x, y,
xend, yend (meant for input into geom_segment()).
}
