% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{find_angle}
\alias{find_angle}
\title{Return angle of rotation of matrix.}
\usage{
find_angle(m)
}
\arguments{
\item{m}{2x2 matrix.}
}
\value{
Double of length 1.
}
\description{
Return angle of rotation of matrix.
}
