% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Create Rotation Matrix}
\usage{
rotation_matrix(theta)
}
\arguments{
\item{theta}{Angle in radians.}
}
\value{
2x2 matrix.
}
\description{
Create Rotation Matrix
}
