% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_matrix.R
\name{seq_matrix_l}
\alias{seq_matrix_l}
\title{Sequence of Matrices, Linear}
\usage{
seq_matrix_l(from, to, ...)
}
\arguments{
\item{from}{Matrix}

\item{to}{Matrix}

\item{...}{Passed into `seq()` function, generally should only input `length.out`.}
}
\value{
List of matrices.
}
\description{
Make sequence of matrices, changing coordinates linearly
}
