% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_matrix.R
\name{seq_matrix_rotate_first}
\alias{seq_matrix_rotate_first}
\title{Sequence of Matrices, polar followed by linear}
\usage{
seq_matrix_rotate_first(from, to, ...)
}
\arguments{
\item{from}{Matrix}

\item{to}{Matrix}

\item{...}{Passed into `seq()` function, generally should only input `length.out`.}
}
\value{
List of matrices.
}
\description{
Generally should not use. Given an individual transformation matrix, return a
combination of polar and linear components. Rotation though happens on a unit
circle.
}
