% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_transform_coordinates.R
\name{transform_df_coords}
\alias{transform_df_coords}
\title{Transform coordinates}
\usage{
transform_df_coords(df, ..., m = diag(length(df)))
}
\arguments{
\item{df}{Dataframe with coordinates.}

\item{...}{Column names where coordinates are (Will be treated as x, y, ...).
Column names not supplied will be unchanged.}

\item{m}{Matrix transformation to apply.}
}
\value{
Dataframe with transformed coordinates.
}
\description{
Given dataframe, column names of coordinates, and a transformation matrix,
return dataframe with transformed coordinates.
}
