% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_transform_coordinates.R
\name{transform_segment}
\alias{transform_segment}
\title{Transform Coordinates of Segment}
\usage{
transform_segment(df, m)
}
\arguments{
\item{df}{Dataframe with column names `x`, `y`, `xend`, `yend`.}

\item{m}{2x2 matrix.}
}
\value{
Dataframe with same column names but coordinates transformed according to `m`.
}
\description{
Helper function to apply `transform_df_coords()` to the dataframe constructed
by `construct_grid()` or similar.
}
